Fixed buggy dark mode toggle
