+++ gatsby-browser.js
@@ -1,18 +1,21 @@
 import React from 'react';
 import { AnimatePresence } from 'framer-motion';
+import { Layout } from './src/components/layout';
 
 export const wrapPageElement = ({ element }) => (
-	<AnimatePresence
-		mode='wait'
-		initial={false}
-		onExitComplete={() => {
-			if (typeof window !== 'undefined') {
-				window.scrollTo({ top: 0 });
-			}
-		}}
-	>
-		{element}
-	</AnimatePresence>
+	<Layout>
+		<AnimatePresence
+			mode='wait'
+			initial={false}
+			onExitComplete={() => {
+				if (typeof window !== 'undefined') {
+					window.scrollTo({ top: 0 });
+				}
+			}}
+		>
+			{element}
+		</AnimatePresence>
+	</Layout>
 );
 
 export const shouldUpdateScroll = () => false;
